+++ src/App.css
@@ -28,3 +28,6 @@ p {
   padding: 0;
 }
 
+button {
+  margin: 0 20px;
+}
